% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{check_pks}
\alias{check_pks}
\title{setup packages for ENCODE work}
\usage{
check_pks(do_install = FALSE)
}
\arguments{
\item{do_install}{logical(1) if TRUE attempt installation}
}
\description{
setup packages for ENCODE work
}
