% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NRpoly}
\alias{NRpoly}
\title{results of quadratic lme fits to accessibility widths in promoters on 17q}
\format{
GRanges
}
\usage{
NRpoly
}
\description{
results of quadratic lme fits to accessibility widths in promoters on 17q
}
\keyword{datasets}
