% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{widScoSE}
\alias{widScoSE}
\title{results of search of p17q for accessibility in promoter regions}
\format{
GRanges
}
\usage{
widScoSE
}
\description{
results of search of p17q for accessibility in promoter regions
}
\keyword{datasets}
